Refactor generic views; add plugins dev documentation
